Update dev on fork

+++ src/Hangfire.SqlServer/IPersistentJobQueueMonitoringApi.cs
@@ -22,11 +22,9 @@ namespace Hangfire.SqlServer
     {
         IEnumerable<string> GetQueues();
 
-        // TODO: Change return type to `IEnumerable<long>` to support `bigint` type.
-        IEnumerable<int> GetEnqueuedJobIds(string queue, int from, int perPage);
+        IEnumerable<long> GetEnqueuedJobIds(string queue, int from, int perPage);
 
-        // TODO: Change return type to `IEnumerable<long>` to support `bigint` type.
-        IEnumerable<int> GetFetchedJobIds(string queue, int from, int perPage);
+        IEnumerable<long> GetFetchedJobIds(string queue, int from, int perPage);
 
         EnqueuedAndFetchedCountDto GetEnqueuedAndFetchedCount(string queue);
     }
